refactor: clean up migrated OAuth model imports and remove unused TokenTransformConfig
