--- registry/src/registry/models/__init__.py
@@ -1,21 +1,0 @@
-from .emus import TokenType
-from .oauth_models import (
-    MCPOAuthFlowMetadata,
-    OAuthClientInformation,
-    OAuthFlow,
-    OAuthMetadata,
-    OAuthProtectedResourceMetadata,
-    OAuthTokens,
-    TokenTransformConfig,
-)
-
-__all__ = [
-    "OAuthTokens",
-    "OAuthClientInformation",
-    "OAuthMetadata",
-    "OAuthProtectedResourceMetadata",
-    "TokenTransformConfig",
-    "MCPOAuthFlowMetadata",
-    "OAuthFlow",
-    "TokenType",
-]
