+++ registry/src/registry/schemas/enums.py
@@ -19,3 +19,11 @@ class ConnectionState(StrEnum):
     PENDING_OAUTH = "pending_oauth"
     ERROR = "error"
     UNKNOWN = "unknown"
+
+
+class TokenType(StrEnum):
+    """Token type enumeration"""
+
+    MCP_OAUTH = "mcp_oauth"  # Access token
+    MCP_OAUTH_REFRESH = "mcp_oauth_refresh"  # Refresh token
+    MCP_OAUTH_CLIENT = "mcp_oauth_client"  # Client credentials (client_id, client_secret)
