+++ registry/src/registry/services/oauth/token_service.py
@@ -5,8 +5,8 @@ from typing import Any
 
 from beanie import PydanticObjectId
 
-from registry.models.emus import TokenType
-from registry.models.oauth_models import OAuthClientInformation, OAuthTokens
+from registry.schemas.enums import TokenType
+from registry.schemas.oauth_sachema import OAuthClientInformation, OAuthTokens
 from registry.services.user_service import user_service
 from registry.utils.crypto_utils import decrypt_auth_fields, encrypt_value
 from registry_pkgs.models import IUser
