+++ registry/tests/unit/services/test_oauth_service.py
@@ -4,8 +4,8 @@ import pytest
 from bson import ObjectId
 
 from registry.auth.oauth import FlowStateManager
-from registry.models.oauth_models import OAuthTokens
 from registry.schemas.enums import OAuthFlowStatus
+from registry.schemas.oauth_sachema import OAuthTokens
 from registry.services.oauth.oauth_service import MCPOAuthService
 from registry.services.oauth.token_service import token_service
 from registry_pkgs.models.extended_mcp_server import ExtendedMCPServer as MCPServerDocument
@@ -269,7 +269,7 @@ class TestMCPOAuthService:
         }
 
         # Mock DCR registration
-        from registry.models.oauth_models import OAuthClientInformation
+        from registry.schemas.oauth_sachema import OAuthClientInformation
 
         mock_client_info = OAuthClientInformation(
             client_id="dcr_registered_client_123",
